separate build functions
